allow VMs to be resumable
There is a new pseudo payload called "resume", which should be enabled
after "lvm", touches /.resume. The rest of the scripts handles this file
and for example do not create/delete the per VM directory.

While the "resume" payload is handy, touching /.resume in the started VM
has the same effect.

It is up to the user to remove that file to finally destroy the VM.