ch2vm: make it more signal save
Now we start to ignore signals in the "critical section", so we don't
interrupt zfs in the middle of some clone action.

If we receive a TERM (e.g., systemctl stop), we ignore that signal, but
remember it. When out of the critical section, we termintate if we
received such a signal.

If not, we still keep ignoring that signal, which is fine, because the
qemu in the subshell does not ignore it and terminates, which finally
terminates ch2vm and calls the EXIT trap.

Having fun with signals in shell! ;-)