add passing of block devices
The "lvm" payload learned some additional options. One of them is
"bd=size". If it is set, a block device (zvol) is generated for this VM
and passed to the VM. The LVM pool will be placed on a PV matching the
passed in block device. "size" has to be a string understood by the zfs
tools. The other option is "thin", and if set, allocates the block
device thinly (zfs create -s ...). As usual these options are passed
separated by ':', as in:
"payloadX;lvm:bd=1g:thin;shell"