vmshed: add zfs integration
This adds zfs integration to vmshed by allowing to new fields in the
json configuration. For VMs, one can set "zfs: true", and in tests, one
can set "needzfs:[]". The later one behaves like "samevms".

while at it, improve vm output:

Every tests prints as a "header" in which VMs the test will be executed
(distribution, kernel, ...). Over the time we added some hacks that the
actual VM executed could have been a different one that actually
printed. For example when "samevms" was set.

Add a function that does these transformations, and then print the
header.