vmshed: add needallplatforms option
This add a new key ('needallplatforms') key, with the usual test string
array. If a test is marked like this, it is executed for every
platform/distribution defined via -vms.

For example LINSTOR can use this to have one test that executes all LVM
commands. And now they can specify that it should run on all platforms,
which helps to protect against regressions in the LVM path (e.g., by
using a too new lvmX command that is not available on every
distribution).